requirements.txt
apache-kafka==2.0.2
apache-nifi==1.15.0
apache-spark==3.1.2
google-cloud-bigquery==2.34.3
cassandra-driver==3.25.0
flask==2.0.2
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
pyspark==3.1.2
pytest==6.2.5
tableau-api-lib==0.1.27
docker==7.1.0
kubernetes==18.20.0